scanner: Break the scanner up into multiple modules
